[game_mechanics] dt is Double and improve controller

+++ src/main/scala/game_mechanics/throw.scala
@@ -15,7 +15,7 @@ class Throw (B:Bunny, T: Waypoint) {
     var pos      = T
 
 /* Update of the position of the throw */
-    def update_pos(dt : Int, dir: Waypoint): Unit = {
+    def update_pos(dt : Double, dir: Waypoint): Unit = {
       val a = this.pos
       this.pos += (this.target.pos - this.pos) * dt * this.speed
       if (((this.pos-this.target.pos) & a) <= 0.0) {
@@ -24,7 +24,7 @@ class Throw (B:Bunny, T: Waypoint) {
     }
 
 /* One step of progress */
-    def progress(dt: Int, dir: Waypoint): Unit= {
+    def progress(dt: Double, dir: Waypoint): Unit= {
       update_pos(dt,dir)
       if (this.target.pos == this.pos) {
         target.takedamage(this.damage)
@@ -33,7 +33,7 @@ class Throw (B:Bunny, T: Waypoint) {
 
     var time_to_touch = 1
 
-    def auto_touch(dt : Int): Unit = {
+    def auto_touch(dt : Double): Unit = {
       if (time_to_touch == 0) {
       this.pos = this.target.pos
     }
